+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/service/DaoOnlyCrudService.java
@@ -21,6 +21,7 @@ import java.util.Objects;
  * @author DwArFeng
  * @since 0.0.1-beta
  */
+@SuppressWarnings("DuplicatedCode")
 public class DaoOnlyCrudService<K extends Key, E extends Entity<K>> implements CrudService<K, E> {
 
     private BaseDao<K, E> dao;
@@ -62,6 +63,9 @@ public class DaoOnlyCrudService<K extends Key, E extends Entity<K>> implements C
     }
 
     private E internalGet(K key) throws Exception {
+        if (!dao.exists(key)) {
+            throw new ServiceException(ServiceExceptionCodes.ENTITY_NOT_EXIST);
+        }
         return dao.get(key);
     }
 
@@ -159,7 +163,6 @@ public class DaoOnlyCrudService<K extends Key, E extends Entity<K>> implements C
         }
     }
 
-    @SuppressWarnings("DuplicatedCode")
     @Override
     public K insertOrUpdate(E element) throws ServiceException {
         try {